+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### ADDED
 * Create dmg installer for Mac build
+* Add a new window to allow user to set their defined man page for a command with hint from the db.
 
 ### CHANGED
+* GUI improvements.
 
 ### DEBUGGED
 * Fix command heightening unable to highlight `&minus;` (a way to represent `-` in html)